+++ handler-annotations/src/main/java/module-info.java
@@ -1,7 +1,7 @@
-module handlerannotations {
+module com.ensarsarajcic.neovim.java.handlerannotations {
     exports com.ensarsarajcic.neovim.java.handler;
     exports com.ensarsarajcic.neovim.java.handler.annotations;
 
-    requires corerpc;
+    requires com.ensarsarajcic.neovim.java.corerpc;
     requires org.slf4j;
 }