+++ reactive-core-rpc/src/main/java/module-info.java
@@ -1,7 +1,7 @@
-module reactivecorerpc {
+module com.ensarsarajcic.neovim.java.reactivecorerpc {
     opens com.ensarsarajcic.neovim.java.corerpc.reactive;
     exports com.ensarsarajcic.neovim.java.corerpc.reactive;
 
-    requires corerpc;
+    requires com.ensarsarajcic.neovim.java.corerpc;
     requires org.slf4j;
 }