+++ testing-helpers/src/main/java/module-info.java
@@ -1,4 +1,4 @@
-module testinghelpers {
+module com.ensarsarajcic.neovim.java.testinghelpers {
     opens com.ensarsarajcic.neovim.java.testing;
     exports com.ensarsarajcic.neovim.java.testing;
 }