+++ unix-socket-connection/src/main/java/module-info.java
@@ -1,7 +1,7 @@
-module unixsocketconnection {
+module com.ensarsarajcic.neovim.java.unixsocketconnection {
     exports com.ensarsarajcic.neovim.java.unix.socket;
 
-    requires corerpc;
+    requires com.ensarsarajcic.neovim.java.corerpc;
     requires ipcsocket;
     requires org.slf4j;
 }